feat: map bottom sheet height 적용해 현재 위치 아이콘 움직이기 (#36)
