+++ src/stores/index.ts
@@ -7,3 +7,4 @@ export { default as useBottomSheetStore } from './common/useBottomSheetStore';
 export { default as useFilterTabsStore } from './event/useFilterTabsStore';
 export { default as useMyLocationStore } from './event/useMyLocationStore';
 export { default as useFilteredEventStore } from './event/useFilteredEventStore';
+export { default as useMapStore } from './event/useMapStore';
