+++ src/utils/getLabelFromValue.ts
@@ -2,15 +2,13 @@ import {
   CATEGORY_OPTIONS,
   PRICE_OPTIONS,
   LOCATION_OPTIONS,
-  SORT_OPTIONS,
 } from '@/constants/event';
 
 const getLabelFromValue = (
   value: string,
-  optionType: 'sort' | 'category' | 'price' | 'location', // 추가해 사용해주세요
+  optionType: 'category' | 'price' | 'location', // 추가해 사용해주세요
 ): string => {
   const optionsMap = {
-    sort: SORT_OPTIONS,
     category: CATEGORY_OPTIONS,
     price: PRICE_OPTIONS,
     location: LOCATION_OPTIONS,
