+++ src/utils/index.ts
@@ -14,4 +14,5 @@ export { default as getLabelFromValue } from './getLabelFromValue';
 export { default as isInRange } from './isInDateRange';
 export { default as isSameDay } from './isSameDay';
 
-export { getLocationPosition } from './location';
+export { default as getCurrentPosition } from './getCurrentPosition';
+export { default as getMarker } from './getMarker';
